Clean up calls
BinanceWallet does not allow the use of custom RPCs and appears to be severely rate limited recently.

In reaction to that we went through and made sure any providers that we create are only turned into signers if necessary (therefore we can use a random RPC instead of relying on the BinanceWallet one)

There was only ~8 offending cases that were actually in use, however a couple of them were often batched in as 10-20 calls at a time, so this should make a big difference.

Other than that, BinanceWallet seems to perform all good, so we can close #852